requirements.txt
websocket-client
numpy
scikit-learn
flask
gunicorn
